+++ src/Raven.CodeAnalysis/Syntax/Quoter/RavenQuoter.cs
@@ -73,9 +73,18 @@ public static class RavenQuoter
 
             w.Write("return ");
             writer.WriteNode(root);
-            w.WriteLine(options.OutputLanguage == RavenQuoterOutputLanguage.Raven
-                ? ".NormalizeWhitespace()"
-                : ".NormalizeWhitespace();");
+            if (options.NormalizeWhitespace)
+            {
+                w.WriteLine(options.OutputLanguage == RavenQuoterOutputLanguage.Raven
+                    ? ".NormalizeWhitespace()"
+                    : ".NormalizeWhitespace();");
+            }
+            else
+            {
+                w.WriteLine(options.OutputLanguage == RavenQuoterOutputLanguage.Raven
+                    ? string.Empty
+                    : ";");
+            }
 
             w.Unindent();
             w.WriteLine("}");
@@ -85,9 +94,18 @@ public static class RavenQuoter
         else
         {
             writer.WriteNode(root);
-            w.WriteLine(options.OutputLanguage == RavenQuoterOutputLanguage.Raven
-                ? ".NormalizeWhitespace()"
-                : ".NormalizeWhitespace();");
+            if (options.NormalizeWhitespace)
+            {
+                w.WriteLine(options.OutputLanguage == RavenQuoterOutputLanguage.Raven
+                    ? ".NormalizeWhitespace()"
+                    : ".NormalizeWhitespace();");
+            }
+            else
+            {
+                w.WriteLine(options.OutputLanguage == RavenQuoterOutputLanguage.Raven
+                    ? string.Empty
+                    : ";");
+            }
         }
 
         return w.ToString();
@@ -169,7 +187,12 @@ public static class RavenQuoter
 
         private void WriteFactoryMethodName(string methodName)
         {
-            if (_options.UseStaticSyntaxFactoryImport)
+            if (_options.FullyQualifyNames)
+            {
+                _w.Write("Raven.CodeAnalysis.Syntax.SyntaxFactory.");
+                _w.Write(methodName);
+            }
+            else if (_options.UseStaticSyntaxFactoryImport)
                 _w.Write(methodName);
             else
             {
@@ -399,7 +422,7 @@ public static class RavenQuoter
 
             if (type.IsEnum)
             {
-                _w.Write(type.Name);
+                WriteTypeName(type);
                 _w.Write(".");
                 _w.Write(Enum.GetName(type, value) ?? value.ToString());
                 return;
@@ -457,14 +480,18 @@ public static class RavenQuoter
                         if (token.IsMissing)
                         {
                             WriteFactoryMethodName("MissingToken");
-                            _w.Write("(SyntaxKind.");
+                            _w.Write("(");
+                            WriteTypeName(typeof(SyntaxKind));
+                            _w.Write(".");
                             _w.Write(token.Kind.ToString());
                             _w.Write(")");
                         }
                         else
                         {
                             WriteFactoryMethodName("Token");
-                            _w.Write("(SyntaxKind.");
+                            _w.Write("(");
+                            WriteTypeName(typeof(SyntaxKind));
+                            _w.Write(".");
                             _w.Write(token.Kind.ToString());
                             _w.Write(")");
                         }
@@ -504,14 +531,18 @@ public static class RavenQuoter
                     if (token.IsMissing)
                     {
                         WriteFactoryMethodName("MissingToken");
-                        _w.Write("(SyntaxKind.");
+                        _w.Write("(");
+                        WriteTypeName(typeof(SyntaxKind));
+                        _w.Write(".");
                         _w.Write(token.Kind.ToString());
                         _w.Write(")");
                     }
                     else
                     {
                         WriteFactoryMethodName("Token");
-                        _w.Write("(SyntaxKind.");
+                        _w.Write("(");
+                        WriteTypeName(typeof(SyntaxKind));
+                        _w.Write(".");
                         _w.Write(token.Kind.ToString());
                         _w.Write(")");
                     }
@@ -597,9 +628,9 @@ public static class RavenQuoter
             if (list.Count == 1)
             {
                 WriteFactoryMethodName("TriviaList");
-                _w.Write("(");
+                _w.Write("([");
                 WriteSingleTrivia(list[0]);
-                _w.Write(")");
+                _w.Write("])");
                 return;
             }
 
@@ -628,7 +659,9 @@ public static class RavenQuoter
         private void WriteSingleTrivia(SyntaxTrivia trivia)
         {
             WriteFactoryMethodName("Trivia");
-            _w.Write("(SyntaxKind.");
+            _w.Write("(");
+            WriteTypeName(typeof(SyntaxKind));
+            _w.Write(".");
             _w.Write(trivia.Kind.ToString());
             _w.Write(", ");
             _w.Write(Literal(trivia.ToString()));
@@ -638,14 +671,13 @@ public static class RavenQuoter
         private void WriteSyntaxList(Type elemType, IEnumerable listEnumerable)
         {
             var items = listEnumerable.Cast<object?>().ToList();
-            var typeName = elemType.Name;
 
             if (items.Count == 0)
             {
-                //_w.Write($"SyntaxList<{typeName}>.Empty");
-
                 WriteFactoryMethodName("List");
-                _w.Write($"<{typeName}>()");
+                _w.Write("<");
+                WriteTypeName(elemType);
+                _w.Write(">()");
 
                 return;
             }
@@ -653,7 +685,9 @@ public static class RavenQuoter
             if (items.Count == 1)
             {
                 WriteFactoryMethodName("SingletonList");
-                _w.Write($"<{typeName}>(");
+                _w.Write("<");
+                WriteTypeName(elemType);
+                _w.Write(">(");
 
                 var hasMoreThanOneChild = items[0] is SyntaxNode sn && sn.DescendantNodes().Any();
 
@@ -675,9 +709,15 @@ public static class RavenQuoter
             }
 
             WriteFactoryMethodName("List");
-            _w.Write($"<{typeName}>(");
+            _w.Write("<");
+            WriteTypeName(elemType);
+            _w.Write(">(");
             if (_options.OutputLanguage == RavenQuoterOutputLanguage.CSharp)
-                _w.Write($"new {typeName}[]");
+            {
+                _w.Write("new ");
+                WriteTypeName(elemType);
+                _w.Write("[]");
+            }
             _w.WriteLine();
             _w.WriteLine(_options.OutputLanguage == RavenQuoterOutputLanguage.Raven ? "[" : "{");
             _w.Indent();
@@ -698,14 +738,13 @@ public static class RavenQuoter
         private void WriteSeparatedSyntaxList(Type elemType, IEnumerable listEnumerable)
         {
             var items = listEnumerable.Cast<object?>().ToList();
-            var typeName = elemType.Name;
 
             if (items.Count == 0)
             {
-                //_w.Write($"SeparateSyntaxSyntax<{typeName}>.Empty");
-
                 WriteFactoryMethodName("SeparatedList");
-                _w.Write($"<{typeName}>()");
+                _w.Write("<");
+                WriteTypeName(elemType);
+                _w.Write(">()");
 
                 return;
             }
@@ -713,7 +752,9 @@ public static class RavenQuoter
             if (items.Count == 1)
             {
                 WriteFactoryMethodName("SingletonSeparatedList");
-                _w.Write($"<{typeName}>(");
+                _w.Write("<");
+                WriteTypeName(elemType);
+                _w.Write(">(");
 
                 var hasMoreThanOneChild = items[0] is SyntaxNode sn && sn.DescendantNodesAndTokens().Count() > 1;
 
@@ -735,9 +776,15 @@ public static class RavenQuoter
             }
 
             WriteFactoryMethodName("SeparatedList");
-            _w.Write($"<{typeName}>(");
+            _w.Write("<");
+            WriteTypeName(elemType);
+            _w.Write(">(");
             if (_options.OutputLanguage == RavenQuoterOutputLanguage.CSharp)
-                _w.Write($"new {typeName}[]");
+            {
+                _w.Write("new ");
+                WriteTypeName(elemType);
+                _w.Write("[]");
+            }
             _w.WriteLine();
             _w.WriteLine(_options.OutputLanguage == RavenQuoterOutputLanguage.Raven ? "[" : "{");
             _w.Indent();
@@ -753,6 +800,13 @@ public static class RavenQuoter
 
             _w.Unindent();
             _w.Write(_options.OutputLanguage == RavenQuoterOutputLanguage.Raven ? "])" : "})");
+        }
+
+        private void WriteTypeName(Type type)
+        {
+            _w.Write(_options.FullyQualifyNames
+                ? type.FullName ?? type.Name
+                : type.Name);
         }
 
         private static bool IsDefaultValue(Type type, object? value)